{"schema":"decodex/commit/1","summary":"Modularize radar artifact validation","authority":"XY-1115"}
