+++ apps/decodex/src/radar/artifact_validation/bundle.rs
@@ -1,0 +1,85 @@
+//! GitHub bundle schema validation.
+
+use serde_json::{Map, Value};
+
+use super::{
+	constants::ANALYSIS_MODES,
+	support::{choices, is_non_empty_string, matches_one_of, non_empty_array, string_field},
+};
+
+pub(super) fn validate_bundle(bundle: &Map<String, Value>, errors: &mut Vec<String>) {
+	if string_field(bundle, "repo").is_none_or(|repo| !repo.contains('/')) {
+		errors.push("repo must be owner/name".into());
+	}
+	if !matches_one_of(bundle.get("analysis_mode"), ANALYSIS_MODES) {
+		errors.push(format!("analysis_mode must be one of {}", choices(ANALYSIS_MODES)));
+	}
+	if !is_non_empty_string(bundle.get("default_branch")) {
+		errors.push("default_branch must be a non-empty string".into());
+	}
+
+	validate_bundle_commits(bundle.get("commits"), errors);
+	validate_bundle_files(bundle.get("files"), errors);
+
+	if string_field(bundle, "analysis_mode") == Some("pr_first") {
+		validate_bundle_pr(bundle.get("primary_pr"), errors);
+	}
+}
+
+pub(super) fn validate_bundle_commits(commits: Option<&Value>, errors: &mut Vec<String>) {
+	let Some(commits) = non_empty_array(commits) else {
+		errors.push("commits must be a non-empty list".into());
+
+		return;
+	};
+
+	for (index, commit) in commits.iter().enumerate() {
+		let Some(commit) = commit.as_object() else {
+			errors.push(format!("commits[{index}] must be an object"));
+
+			continue;
+		};
+
+		for field in ["sha", "message", "url"] {
+			if !is_non_empty_string(commit.get(field)) {
+				errors.push(format!("commits[{index}].{field} must be a non-empty string"));
+			}
+		}
+	}
+}
+
+pub(super) fn validate_bundle_files(files: Option<&Value>, errors: &mut Vec<String>) {
+	let Some(files) = non_empty_array(files) else {
+		errors.push("files must be a non-empty list".into());
+
+		return;
+	};
+
+	for (index, item) in files.iter().enumerate() {
+		let Some(item) = item.as_object() else {
+			errors.push(format!("files[{index}] must be an object"));
+
+			continue;
+		};
+
+		for field in ["path", "status", "additions", "deletions"] {
+			if !item.contains_key(field) {
+				errors.push(format!("files[{index}].{field} is required"));
+			}
+		}
+	}
+}
+
+pub(super) fn validate_bundle_pr(primary_pr: Option<&Value>, errors: &mut Vec<String>) {
+	let Some(primary_pr) = primary_pr.and_then(Value::as_object) else {
+		errors.push("primary_pr is required when analysis_mode is pr_first".into());
+
+		return;
+	};
+
+	for field in ["number", "title", "body", "state", "labels", "url"] {
+		if !primary_pr.contains_key(field) {
+			errors.push(format!("primary_pr.{field} is required"));
+		}
+	}
+}
